requirements.txt
alpaca-trade-api
pandas
numpy
matplotlib
seaborn
python-dotenv
